Restored the project to original state before I tried adding arrays.
There seems to be a major issue caused by a black friday alteration to the amazon site that affects exclusively tech product searches.

This results in the image object failing and causes the extract record function to either loop endlessly (when no image or image_parent is returned) or to corrupt the csv by replacing the entire row with the comma seperated return value of image AttributeError (When a return value for image is specified).